+++ src/main/java/com/mdh/devtable/reservation/domain/SeatStatus.java
@@ -1,4 +1,4 @@
-package com.mdh.devtable.reservation;
+package com.mdh.devtable.reservation.domain;
 
 public enum SeatStatus {
     AVAILABLE,
