+++ src/main/java/com/mdh/devtable/reservation/domain/SeatType.java
@@ -1,4 +1,4 @@
-package com.mdh.devtable.reservation;
+package com.mdh.devtable.reservation.domain;
 
 public enum SeatType {
     ROOM,
