+++ src/main/java/com/mdh/devtable/reservation/infra/persistence/ShopReservationDateTimeSeatRepository.java
@@ -1,6 +1,6 @@
 package com.mdh.devtable.reservation.infra.persistence;
 
-import com.mdh.devtable.reservation.ShopReservationDateTimeSeat;
+import com.mdh.devtable.reservation.domain.ShopReservationDateTimeSeat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ShopReservationDateTimeSeatRepository extends JpaRepository<ShopReservationDateTimeSeat, Long> {
